+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ### v2.3 - The Open Source Update
 ###### April ??, 2025
-*   [!!!!] This project - dubbed the Speedrun Leaderboard Creator (SRLC) - is now open source! For more information on this, please check out the project's github here: [https://github.com/ThePackle/SRLC/](https://github.com/ThePackle/SRLC/)
+*   [!!!!] This project is now open source! For more information on this, please check out the project's GitHub here: [https://github.com/ThePackle/SRLC/](https://github.com/ThePackle/SRLC/)
   
 *   Added `bluesky` to the `Players` model.
     *   As a note, this is minor change. If you want your Bluesky account added to your profile, contact Packle. A later update will move to SRC's v2 endpoint, which can return Bluesky links; otherwise this must be done manually.
@@ -10,13 +10,16 @@
 *   Added Bluesky as an environmental variable to the project that can display on the navbar.
 *   Added support for a run's current run verification status from SRC.
     *   Runs will be cached from now on when they are awaiting approval. They will be removed if rejected.
-*   Added support for an optional archived videos field to `MainRuns` and `ILRuns` models.
+*   Added support for an optional archived videos field to `Runs` model.
 *   Added quick-links to Twitch, YouTube, and/or Archived Video to player profiles, leaderboards, and the main page. [#19](https://github.com/ThePackle/SRLC/issues/19)
 *   Added "Run History" for all players. By navigating to `/player/<NAME>/history/`, you will be able to see ALL current and obsolete speedruns (ordered by date).
     *   Quick-links to Twitch, YouTube, and/or Archived Videos are also here!
 *   Added more alt and title tags through the site for accessibility.
 *   Added a new panel on the main page to show who is actively streaming.
     *   Also added the `/live/` endpoint, more information below.
+*   Added the new `Runs` model to replace `MainRuns` and `ILRuns`.
+    *   It was frustrating dealing with duplicate code for two nearly identical models (`MainRuns` has support for 2 players; `ILRuns` has support for Levels). This update forced me to change this to a combined model - had to get rid of dumb tech debt!
+    *   Each runtype (`main` and `il`) have been given new QuerySet options to better separate them quickly.
   
   
 *   Changed the front page to display tied world records on the same table row. [#12](https://github.com/ThePackle/SRLC/issues/12)
@@ -42,6 +45,8 @@
 *   Fixed an issue where, if a game's timing is set to LRT and no RTA was submitted, it would appear as "0m 00s" on the leaderboard.
     *   Only obsolete runs were really affected, so this is more for the "Runs History" page for players.
   
+
+*   Removed `MainRuns` and `ILRuns` models.
 *   Removed Twitter from the navbar.
     *   If you want to use this in your projects, uncomment the code in navbar.html.
 *   Removed `Location` field from the Players model.